Reorganized rides routes to format specified by rider dashboard issue
