+++ frontend/src/App.js
@@ -4,9 +4,9 @@ import ProfilePage from "main/pages/ProfilePage";
 import AdminUsersPage from "main/pages/AdminUsersPage";
 import PageNotFound from "main/pages/PageNotFound";
 
-import RideRequestCreatePage from "main/pages/Ride/RideRequestCreatePage";
-import RideRequestEditPage from "main/pages/Ride/RideRequestEditPage";
-import RideRequestIndexPage from "main/pages/Ride/RideRequestIndexPage";
+import RideRequestCreatePage from "main/pages/Rider/RideRequestCreatePage";
+import RideRequestEditPage from "main/pages/Rider/RideRequestEditPage";
+import RideRequestIndexPage from "main/pages/Rider/RideRequestIndexPage";
 import ShiftPage from "main/pages/ShiftPage";
 
 
@@ -31,13 +31,13 @@ function App() {
           hasRole(currentUser, "ROLE_ADMIN") && <Route exact path="/admin/users" element={<AdminUsersPage />} />
         }
         {
-          (hasRole(currentUser, "ROLE_ADMIN") || hasRole(currentUser, "ROLE_DRIVER") || hasRole(currentUser, "ROLE_RIDER") )&& <Route exact path="/ride/" element={<RideRequestIndexPage />} />
+          (hasRole(currentUser, "ROLE_ADMIN") || hasRole(currentUser, "ROLE_DRIVER") || hasRole(currentUser, "ROLE_RIDER") )&& <Route exact path="/rider/" element={<RideRequestIndexPage />} />
         }
         {
-          (hasRole(currentUser, "ROLE_RIDER") || hasRole(currentUser, "ROLE_ADMIN"))  && <Route exact path="/ride/create" element={<RideRequestCreatePage />} />
+          (hasRole(currentUser, "ROLE_RIDER") || hasRole(currentUser, "ROLE_ADMIN"))  && <Route exact path="/rider/ride/create" element={<RideRequestCreatePage />} />
         }
         {
-          (hasRole(currentUser, "ROLE_ADMIN")  || hasRole(currentUser, "ROLE_RIDER") )&& <Route exact path="/ride/edit/:id" element={<RideRequestEditPage />} />
+          (hasRole(currentUser, "ROLE_ADMIN")  || hasRole(currentUser, "ROLE_RIDER") )&& <Route exact path="/rider/ride/edit/:id" element={<RideRequestEditPage />} />
         }
         {
           hasRole(currentUser, "ROLE_ADMIN") && <Route exact path="/shift/list" element={<ShiftPage />} />
