+++ frontend/src/main/components/Nav/AppNavbar.js
@@ -12,13 +12,13 @@ function isParticipant(currentUser) {
   );
 }
 
-function createRideRequest(currentUser) {
-  if (hasRole(currentUser, "ROLE_RIDER") || hasRole(currentUser, "ROLE_ADMIN")) {
-    return (<NavDropdown.Item data-testid="appnavbar-ride-create-dropdown" as={Link} to="/ride/create">Request Ride</NavDropdown.Item>)
-  } else {
-    return (<div data-testid="NO-appnavbar-ride-create-dropdown"></div>)
-  }
-}
+// function createRideRequest(currentUser) {
+//   if (hasRole(currentUser, "ROLE_RIDER") || hasRole(currentUser, "ROLE_ADMIN")) {
+//     return (<NavDropdown.Item data-testid="appnavbar-ride-create-dropdown" as={Link} to="/rider/ride/create">Request Ride</NavDropdown.Item>)
+//   } else {
+//     return (<div data-testid="NO-appnavbar-ride-create-dropdown"></div>)
+//   }
+// }
 
 export default function AppNavbar({ currentUser, systemInfo, doLogout, currentUrl = window.location.href }) {
   const styles = {
@@ -82,10 +82,11 @@ export default function AppNavbar({ currentUser, systemInfo, doLogout, currentUr
               }
               {
                 isParticipant(currentUser) && (
-                  <NavDropdown title="Ride Request" id="appnavbar-ride-dropdown" data-testid="appnavbar-ride-dropdown" >
-                    <NavDropdown.Item data-testid="appnavbar-ride-dropdown-rides" as={Link} to="/ride/">Rides</NavDropdown.Item>
-                    { createRideRequest(currentUser) }
-                  </NavDropdown>
+                  // <NavDropdown title="Rider" id="appnavbar-ride-dropdown" data-testid="appnavbar-ride-dropdown" >
+                  //   <NavDropdown.Item data-testid="appnavbar-ride-dropdown-rides" as={Link} to="/rider/">Rider Dashboard</NavDropdown.Item>
+                  //   { createRideRequest(currentUser) }
+                  // </NavDropdown>
+                  <Nav.Link href="/rider/">Rider</Nav.Link>
                 )
               }
             </Nav>
