+++ frontend/src/main/pages/Rider/RideRequestIndexPage.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { useBackend } from 'main/utils/useBackend';
 
 import BasicLayout from "main/layouts/BasicLayout/BasicLayout";
-import RideTable from 'main/components/Ride/RideTable';
+import RideTable from 'main/components/Rider/RideTable';
 import {useCurrentUser } from 'main/utils/currentUser'
 import Button from 'react-bootstrap/Button';
 import { Link } from 'react-router-dom';
@@ -23,11 +23,11 @@ export default function RideRequestIndexPage() {
   return (
     <BasicLayout>
       <div className="pt-2">
-        <Button style={{ float: "right" }} as={Link} to="/ride/create">
+        <Button style={{ float: "right" }} as={Link} to="/rider/ride/create">
           Create Ride Request
         </Button>
         <h1>Ride Requests</h1>
-        <RideTable ride={rides} currentUser={currentUser} />
+        <RideTable rides={rides} currentUser={currentUser} />
       </div>
     </BasicLayout>
   )
