+++ frontend/src/stories/pages/Ride/RideRequestCreatePage.stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import RideRequestCreatePage from 'main/pages/Ride/RideRequestCreatePage';
+import RideRequestCreatePage from 'main/pages/Rider/RideRequestCreatePage';
 
 export default {
     title: 'pages/Ride/RideRequestCreatePage',
