+++ frontend/src/tests/pages/Rider/RideRequestEditPage.test.js
@@ -1,7 +1,7 @@
 import { fireEvent, render, waitFor } from "@testing-library/react";
 import { QueryClient, QueryClientProvider } from "react-query";
 import { MemoryRouter } from "react-router-dom";
-import RideRequestEditPage from "main/pages/Ride/RideRequestEditPage";
+import RideRequestEditPage from "main/pages/Rider/RideRequestEditPage";
 
 import { apiCurrentUserFixtures } from "fixtures/currentUserFixtures";
 import { systemInfoFixtures } from "fixtures/systemInfoFixtures";
